Fix form errors on the Onboarding pages
